Split out artifact for javaagent SPI.
