+++ javaagent-exporters/logging/logging.gradle
@@ -8,7 +8,7 @@ apply from: "$rootDir/gradle/publish.gradle"
 archivesBaseName = 'javaagent-exporters-logging'
 
 dependencies {
-  compileOnly project(':javaagent-tooling')
+  compileOnly project(':javaagent-spi')
 
   annotationProcessor deps.autoservice
   compileOnly deps.autoservice
