+++ javaagent-exporters/otlp/otlp.gradle
@@ -8,7 +8,7 @@ apply from: "$rootDir/gradle/publish.gradle"
 archivesBaseName = 'javaagent-exporters-otlp'
 
 dependencies {
-  compileOnly(project(':javaagent-tooling'))
+  compileOnly(project(':javaagent-spi'))
   compileOnly(deps.opentelemetrySdk)
 
   annotationProcessor deps.autoservice
