+++ javaagent-exporters/otlp/src/main/java/io/opentelemetry/javaagent/exporters/otlp/OtlpSpanExporterFactory.java
@@ -18,8 +18,8 @@ package io.opentelemetry.javaagent.exporters.otlp;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.exporters.otlp.OtlpGrpcSpanExporter;
-import io.opentelemetry.javaagent.tooling.exporter.ExporterConfig;
-import io.opentelemetry.javaagent.tooling.exporter.SpanExporterFactory;
+import io.opentelemetry.javaagent.spi.exporter.ExporterConfig;
+import io.opentelemetry.javaagent.spi.exporter.SpanExporterFactory;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
 
 @AutoService(SpanExporterFactory.class)
