+++ javaagent-exporters/zipkin/zipkin.gradle
@@ -8,7 +8,7 @@ apply from: "$rootDir/gradle/publish.gradle"
 archivesBaseName = 'javaagent-exporters-zipkin'
 
 dependencies {
-  compileOnly(project(':javaagent-tooling'))
+  compileOnly(project(':javaagent-spi'))
   compileOnly(deps.opentelemetrySdk)
 
   annotationProcessor deps.autoservice
