+++ javaagent-spi/javaagent-spi.gradle
@@ -1,0 +1,6 @@
+apply from: "$rootDir/gradle/java.gradle"
+apply from: "$rootDir/gradle/publish.gradle"
+
+dependencies {
+  compileOnly deps.opentelemetrySdk
+}
