+++ javaagent-tooling/javaagent-tooling.gradle
@@ -12,6 +12,7 @@ configurations {
 
 dependencies {
   implementation project(':javaagent-bootstrap')
+  implementation project(':javaagent-spi')
   implementation project(':auto-api')
   implementation project(':instrumentation-api')
