+++ settings.gradle
@@ -37,6 +37,7 @@ include ':opentelemetry-api-beta-shaded-for-instrumenting'
 include ':opentelemetry-auto-annotations-beta-shaded-for-instrumenting'
 include ':opentelemetry-sdk-beta-shaded-for-instrumenting'
 include ':javaagent-bootstrap'
+include ':javaagent-spi'
 include ':javaagent-tooling'
 include ':javaagent'
 include ':load-generator'
